refactor shared SurveySelector logic to a new component
