+++ src/pages/Survey.js
@@ -1,13 +1,13 @@
-import { format, parseISO, startOfWeek } from 'date-fns';
 import PropTypes from 'prop-types';
 import React, { Component } from 'react';
-import { Container, Row, Col, FormControl } from 'react-bootstrap';
+import { Container, Row, Col } from 'react-bootstrap';
 import { Helmet } from 'react-helmet';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import * as Survey from 'survey-react';
 
 import Surveys from 'data/surveys';
+import SurveySelector from 'components/SurveySelector/SurveySelector';
 import { actions as appActions } from 'reducers/application';
 import { getSelectedSurvey, getSurveys } from 'selectors/application';
 
@@ -39,7 +39,6 @@ export class SurveyPage extends Component {
     };
     this.surveys = [];
     this.onComplete = this.onComplete.bind(this);
-    this.handleSurveyChange = this.handleSurveyChange.bind(this);
   }
 
   async componentDidMount() {
@@ -63,54 +62,6 @@ export class SurveyPage extends Component {
     }
 
     actions.submitSurvey(cookieName, valuesHash);
-  }
-
-  handleSurveyChange(event) {
-    const { actions } = this.props;
-    const {
-      target: { value }
-    } = event;
-
-    if (!value) {
-      return;
-    }
-
-    actions.loadSurvey(value);
-  }
-
-  get surveyOptions() {
-    return Surveys.map(survey => {
-      if (survey.visible) {
-        return null;
-      }
-
-      const surveyDate = format(
-        startOfWeek(parseISO(survey.id.replace('fotw-', ''))),
-        'MMM do'
-      );
-
-      return (
-        <option value={survey.id} key={survey.id}>
-          {survey.name || survey.id} - Week of {surveyDate}
-        </option>
-      );
-    });
-  }
-
-  get surveySelector() {
-    const { selectedSurvey } = this.props;
-
-    return (
-      <FormControl
-        as="select"
-        defaultValue={selectedSurvey?.id}
-        onChange={this.handleSurveyChange}
-        className="mb-2"
-      >
-        <option value="">Select a survey</option>
-        {this.surveyOptions}
-      </FormControl>
-    );
   }
 
   get survey() {
@@ -144,7 +95,7 @@ export class SurveyPage extends Component {
         <Row>
           <Col>
             <h1>Active Polls</h1>
-            {this.surveySelector}
+            <SurveySelector showVisible={false} />
             {this.survey}
           </Col>
         </Row>
